+++ crates/nu-command/src/commands.rs
@@ -70,7 +70,6 @@ pub(crate) mod histogram;
 pub(crate) mod history;
 pub(crate) mod if_;
 pub(crate) mod insert;
-pub(crate) mod int_;
 pub(crate) mod into_int;
 pub(crate) mod keep;
 pub(crate) mod last;
@@ -213,7 +212,6 @@ pub(crate) use help::Help;
 pub(crate) use histogram::Histogram;
 pub(crate) use history::History;
 pub(crate) use insert::Command as Insert;
-pub(crate) use int_::Int_;
 pub(crate) use into_int::IntoInt;
 pub(crate) use keep::{Keep, KeepUntil, KeepWhile};
 pub(crate) use last::Last;
