+++ crates/nu-command/src/commands/default_context.rs
@@ -121,7 +121,6 @@ pub fn create_default_context(interactive: bool) -> Result<EvaluationContext, Bo
             whole_stream_command(Get),
             whole_stream_command(Update),
             whole_stream_command(Insert),
-            whole_stream_command(Int_),
             whole_stream_command(IntoInt),
             whole_stream_command(SplitBy),
             // Row manipulation
